[core] Privatize vector_tile.hpp contents
